bt_bulk.c: In function '__wt_db_bulk_load': bt_bulk.c:71:3: warning: negative integer implicitly converted to unsigned type bt_bulk.c: In function '__wt_bulk_var': bt_bulk.c:525:4: warning: conversion to 'uint32_t' from 'long long unsigned int' may alter its value bt_bulk.c:556:3: warning: conversion to 'uint32_t' from 'long long unsigned int' may alter its value bt_bulk.c: In function '__wt_bulk_dup_offpage': bt_bulk.c:773:3: warning: conversion to 'uint32_t' from 'long long unsigned int' may alter its value bt_bulk.c: In function '__wt_bulk_promote': bt_bulk.c:1004:3: warning: conversion to 'uint32_t' from 'int' may change the sign of the result bt_bulk.c:1102:3: warning: conversion to 'uint32_t' from 'long long unsigned int' may alter its value bt_bulk.c:1116:3: warning: conversion to 'uint32_t' from 'long long unsigned int' may alter its value bt_bulk.c: In function '__wt_item_build_data': bt_bulk.c:1225:3: warning: negative integer implicitly converted to unsigned type bt_bulk.c: In function '__wt_bulk_ovfl_copy': bt_bulk.c:1298:2: warning: conversion to 'uint32_t' from 'long long unsigned int' may alter its value bt_bulk.c: In function '__wt_bulk_ovfl_write': bt_bulk.c:1342:2: warning: conversion to 'uint32_t' from 'long long unsigned int' may alter its value

+++ src/btree/bt_bulk.c
@@ -1001,8 +1001,8 @@ split:		switch (dsk->type) {
 		WT_ILLEGAL_FORMAT(db);
 		}
 
-		WT_ERR(__wt_bulk_scratch_page(
-		    toc, db->intlmin, type, dsk->level + 1, &next, &next_tmp));
+		WT_ERR(__wt_bulk_scratch_page(toc, db->intlmin,
+		    type, (uint32_t)dsk->level + 1, &next, &next_tmp));
 		__wt_set_ff_and_sa_from_offset(next,
 		    WT_PAGE_BYTE(next), &next_first_free, &next_space_avail);
